+++ src/components/end-user/achievement/components/DesktopVerticalLine.tsx
@@ -1,8 +1,8 @@
 // UI
 import { Box } from '@chakra-ui/react'
 
-// Utils
-import { getColor } from '@components/end-user/achievement/helpers'
+// Hooks
+import useGetColor from '../useGetColor'
 
 // Types
 import { AchievementType } from '../types'
@@ -13,18 +13,22 @@ export interface IDesktopVerticalLine {
   index: number
 }
 
-const DesktopVerticalLine = ({ index, achievementsArray, achievement }: IDesktopVerticalLine) => (
-  <Box
-    bg={index === achievementsArray.length - 1 ? 'brand.neutral3' : getColor(achievement)}
-    className="desktop-vertical-line"
-    display="block"
-    h="160px"
-    left="39.5px"
-    position="absolute"
-    top="32px"
-    w="1px"
-    zIndex="0"
-  />
-)
+const DesktopVerticalLine = ({ index, achievementsArray, achievement }: IDesktopVerticalLine) => {
+  const { getColor } = useGetColor()
+
+  return (
+    <Box
+      bg={index === achievementsArray.length - 1 ? 'brand.neutral3' : getColor(achievement)}
+      className="desktop-vertical-line"
+      display="block"
+      h="160px"
+      left="39.5px"
+      position="absolute"
+      top="32px"
+      w="1px"
+      zIndex="0"
+    />
+  )
+}
 
 export default DesktopVerticalLine
