+++ src/components/end-user/achievement/components/MobileVerticalTop.tsx
@@ -1,8 +1,8 @@
 // UI
 import { Box } from '@chakra-ui/react'
 
-// Utils
-import { getColor } from '@components/end-user/achievement/helpers'
+// Hooks
+import useGetColor from '../useGetColor'
 
 // Types
 import { IDesktopVerticalLine } from './DesktopVerticalLine'
@@ -12,23 +12,27 @@ import { achievementsArray } from '../constants'
 
 interface IMobileVerticalTop extends Omit<IDesktopVerticalLine, 'achievementsArray'> {}
 
-const MobileVerticalTop = ({ index, achievement }: IMobileVerticalTop) => (
-  <Box
-    alignItems="center"
-    className="mobile-middle-vertical-line-top"
-    display="flex"
-    justifyContent="center"
-    position="relative"
-    w="100%"
-    zIndex="-1"
-  >
+const MobileVerticalTop = ({ index, achievement }: IMobileVerticalTop) => {
+  const { getColor } = useGetColor()
+
+  return (
     <Box
-      bg={index === achievementsArray.length - 1 ? 'brand.neutral3' : getColor(achievement)}
-      h="160px"
-      position="absolute"
-      w="1px"
-    />
-  </Box>
-)
+      alignItems="center"
+      className="mobile-middle-vertical-line-top"
+      display="flex"
+      justifyContent="center"
+      position="relative"
+      w="100%"
+      zIndex="-1"
+    >
+      <Box
+        bg={index === achievementsArray.length - 1 ? 'brand.neutral3' : getColor(achievement)}
+        h="160px"
+        position="absolute"
+        w="1px"
+      />
+    </Box>
+  )
+}
 
 export default MobileVerticalTop
